requirements.txt
flask==3.0.0
flask-cors==4.0.0
qdrant-client==1.7.0
torch==2.1.0
transformers==4.36.0
pillow==10.1.0
numpy==1.24.3
gunicorn==21.2.0
sentencepiece==0.1.99
chonkie==0.1.1
